test(asana.users): separate pagination projection and resume checks
